backend/requirements.txt
flask==2.2.5
werkzeug==2.2.3
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
tensorflow>=2.10.0,<3.0.0
